Refactor changes from PR - improve file structure for contact middleware and route
